isl_coalesce.c: contains: always return equalities back to original state
If the negated form of an equality is not satisfied by the tableau,
then we would return 0 immediately without first reverting the failing
equality back to its original state.  This is important to be able to match
the equality status bits with the right sign of the equality, in particular
if we only want to wrap the non-valid side.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 14c45f1958ed5df54d7a5239935cc6da9133aa9a