generalize isl_multi_*_gist and isl_multi_*_intersect_domain
In particular, parametrize the domain of the elements of the multi expression
in these functions.  For all multi expressions that are currently defined,
or at least those with a gist and an intersect_domain function, the
domains of the elements are sets.  For an isl_multi_union_pw_aff,
which will be introduced next, the domains of these elements are
union sets.  We therefore can no longer assume that they will always
be sets.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 629f994c1271432ec0644a23f23dd29cf6ae3b61