isl_coalesce.c: check_facets: mark tableau rational before subset tests
In theory, it is possible for a constraint to be considered to be a cut
constraint for an entire basic map while it is considered to be a valid
constraint for a facet of the basic map, even though it cuts off the same
rational points because the validity tests on integer tableaus are
somewhat relaxed without any firm guarantee as to how much they are
relaxed.  Mark the tableau as rational to avoid this theoretical possibility,
which could lead to the wrong conclusions.
It's not clear if this could ever happen in practice, but it's better
to be safe.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: bc34b33ab0b9d9b8916b226539c96539b8b8ab16