isl_tab.c: drop_col: drop column before dropping corresponding variable
Since c557799 (add isl_tab_insert_var, Fri Oct 10 11:44:09 2014 +0200),
we allow the insertion of variables in the tableau at positions other
than the final position.  This requires all subsequent variables to
be moved up.  During an undo, these changes need to be undone,
but they need to be postponed until the colum is removed since
the column may need to be moved into the last position and
this requires the positions of the variables involved, including
the variable that is about to be removed.

Reported-by: Sven Wuytack <Sven.Wuytack@synopsys.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 5e2bd4b64083284477ba37fcf47d45ad94748223