add isl_schedule_constraints_set_context
This allows the user to specify additional constraints on the parameters
that may be assumed during the construction of the schedule.

Without this separate context, a user would have to include it
in the domain, but ever since 16f65ef
(isl_schedule_constraints_compute_schedule: produce schedule tree based
schedule, Sun Jul 28 18:40:18 2013 +0200), the domain is also used
in the root of the returned schedule tree and this domain should
not include any extra constraints on the parameters as that may
result in a needless increase in the number of disjuncts.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: f457c93a9511d425b249fcd29b9c876c215b1e83